+++ app/javascript/components/student/TracksList/List/Track/ProgressBar.jsx
