+++ app/javascript/components/student/index.js
@@ -1,1 +1,1 @@
-export { TrackList } from './TrackList'
+export { TracksList } from './TracksList'
