+++ config/routes.rb
@@ -54,7 +54,7 @@ Rails.application.routes.draw do
     namespace :test do
       namespace :components do
         namespace :student do
-          resource :track_list, only: [:show], controller: "track_list" do
+          resource :tracks_list, only: [:show], controller: "tracks_list" do
             member do
               get 'tracks'
             end
